requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.0.0
Flask>=2.0.0
requests>=2.0.0
